some tests for the home page
