+++ chess/src/Components/play.css
@@ -9,3 +9,32 @@
 .chat-wrapper {
   border: 4px solid black;
 }
+
+.chat {
+  border: 4px #7a9e9f;
+  width: 20%;
+  float: right;
+}
+
+input[type="submit"] {
+  width: 60%;
+  background-color: #7a9e9f;
+  color: white;
+  padding: 8px 10px;
+  margin: 8px 0;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+input[type="textarea"],
+select {
+  width: 80%;
+
+  padding: 8px 20px;
+  margin: 8px 0;
+  display: inline-block;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  box-sizing: border-box;
+}
